Add: RolesPermissions, Create Siswa, Presensi
Penambahan SweetAlert dan Beberapa Perbaikan di BackEnd

+++ routes/admin.php
@@ -22,18 +22,18 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Siswa
     Route::get('/siswa', [SiswaController::class, 'index'])->name('siswa');
-    Route::name('siswa.')->prefix('siswa.')->group(function () {
-        Route::get('/create', [SiswaController::class, 'create'])->name('create');
-        Route::post('/store', [SiswaController::class, 'store'])->name('store');
-        Route::get('/{siswa}', [SiswaController::class, 'show'])->name('show');
-        Route::get('/{siswa}/edit', [SiswaController::class, 'edit'])->name('edit');
-        Route::put('/{siswa}', [SiswaController::class, 'update'])->name('update');
-        Route::delete('/{siswa}', [SiswaController::class, 'destroy'])->name('destroy');
+    Route::name('siswa.')->prefix('siswa')->group(function () {
+        Route::get('/create', [SiswaController::class, 'create'])->middleware(['permission:tambah-siswa'])->name('create');
+        Route::post('/store', [SiswaController::class, 'store'])->middleware(['permission:tambah-siswa'])->name('store');
+        Route::get('/{siswa}', [SiswaController::class, 'show'])->middleware(['permission:lihat-siswa'])->name('show');
+        Route::get('/{siswa}/edit', [SiswaController::class, 'edit'])->middleware(['permission:edit-siswa'])->name('edit');
+        Route::put('/{siswa}', [SiswaController::class, 'update'])->middleware(['permission:edit-siswa'])->name('update');
+        Route::delete('/{siswa}', [SiswaController::class, 'destroy'])->middleware(['permission:hapus-siswa'])->name('destroy');
     });
 
     // Tahun Ajaran
     Route::get('/tahun-ajaran', [TahunAjaranController::class, 'index'])->name('tahun-ajaran');
-    Route::name('tahun-ajaran.')->prefix('tahun-ajaran.')->group(function () {
+    Route::name('tahun-ajaran.')->prefix('tahun-ajaran')->group(function () {
         Route::get('/create', [TahunAjaranController::class, 'create'])->name('create');
         Route::post('/store', [TahunAjaranController::class, 'store'])->name('store');
         Route::get('/{tahunAjaran}', [TahunAjaranController::class, 'show'])->name('show');
@@ -44,7 +44,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Jurusan
     Route::get('/jurusan', [JurusanController::class, 'index'])->name('jurusan');
-    Route::name('jurusan.')->prefix('jurusan.')->group(function () {
+    Route::name('jurusan.')->prefix('jurusan')->group(function () {
         Route::get('/create', [JurusanController::class, 'create'])->name('create');
         Route::post('/store', [JurusanController::class, 'store'])->name('store');
         Route::get('/{jurusan}', [JurusanController::class, 'show'])->name('show');
@@ -55,7 +55,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Kelas
     Route::get('/kelas', [KelasController::class, 'index'])->name('kelas');
-    Route::name('kelas.')->prefix('kelas.')->group(function () {
+    Route::name('kelas.')->prefix('kelas')->group(function () {
         Route::get('/create', [KelasController::class, 'create'])->name('create');
         Route::post('/store', [KelasController::class, 'store'])->name('store');
         Route::get('/{kelas}', [KelasController::class, 'show'])->name('show');
@@ -67,7 +67,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
     // Pengguna 
     Route::get('/pengguna', [PenggunaController::class, 'index'])->name('pengguna');
     Route::get('users-export', [PenggunaController::class, 'export'])->name('pengguna.export');
-    Route::name('pengguna.')->prefix('pengguna.')->group(function () {
+    Route::name('pengguna.')->prefix('pengguna')->group(function () {
         Route::get('/create', [PenggunaController::class, 'create'])->name('create');
         Route::post('/store', [PenggunaController::class, 'store'])->name('store');
         Route::get('/{pengguna}', [PenggunaController::class, 'show'])->name('show');
@@ -78,7 +78,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Peringkat
     Route::get('/peringkat', [PeringkatController::class, 'index'])->name('peringkat');
-    Route::name('peringkat.')->prefix('peringkat.')->group(function () {
+    Route::name('peringkat.')->prefix('peringkat')->group(function () {
         Route::get('/create', [PeringkatController::class, 'create'])->name('create');
         Route::post('/store', [PeringkatController::class, 'store'])->name('store');
         Route::get('/{peringkat}', [PeringkatController::class, 'show'])->name('show');
@@ -89,9 +89,10 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Presensi
     Route::get('/presensi', [PresensiController::class, 'index'])->name('presensi');
-    Route::name('presensi.')->prefix('presensi.')->group(function () {
+    Route::name('presensi.')->prefix('presensi')->group(function () {
         Route::get('/create', [PresensiController::class, 'create'])->name('create');
         Route::post('/store', [PresensiController::class, 'store'])->name('store');
+        Route::get('/autofill', [PresensiController::class, 'autofill'])->name('autofill');
         Route::get('/{presensi}', [PresensiController::class, 'show'])->name('show');
         Route::get('/{presensi}/edit', [PresensiController::class, 'edit'])->name('edit');
         Route::put('/{presensi}', [PresensiController::class, 'update'])->name('update');
@@ -100,7 +101,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Profil
     Route::get('/profil', [ProfilController::class, 'index'])->name('profil');
-    Route::name('profil.')->prefix('profil.')->group(function () {
+    Route::name('profil.')->prefix('profil')->group(function () {
         Route::get('/{profil}', [ProfilController::class, 'show'])->name('show');
         Route::get('/{profil}/edit', [ProfilController::class, 'edit'])->name('edit');
         Route::put('/{profil}', [ProfilController::class, 'update'])->name('update');
@@ -109,7 +110,7 @@ Route::name('admin.')->prefix('admin')->middleware(['auth:admin', 'auth.session'
 
     // Laporan
     Route::get('/laporan', [LaporanController::class, 'index'])->name('laporan');
-    Route::name('laporan.')->prefix('laporan.')->group(function () {
+    Route::name('laporan.')->prefix('laporan')->group(function () {
         Route::get('/{laporan}', [LaporanController::class, 'show'])->name('show');
         Route::get('/{laporan}/edit', [LaporanController::class, 'edit'])->name('edit');
         Route::put('/{laporan}', [LaporanController::class, 'update'])->name('update');
